+++ src/Proto.Cluster/PubSubContracts.proto
@@ -1,0 +1,72 @@
+syntax = "proto3";
+package cluster.pubsub;
+option csharp_namespace = "Proto.Cluster.PubSub";
+import "Proto.Actor/Protos.proto";
+import "ClusterContracts.proto";
+
+message SubscriberIdentity {
+  oneof Identity {
+    actor.PID pid = 1;
+    cluster.ClusterIdentity cluster_identity = 2;
+  }
+}
+
+message Subscribers {
+  repeated SubscriberIdentity subscribers = 1;
+}
+
+message SubscribeRequest {
+  SubscriberIdentity subscriber = 1;
+}
+
+message SubscribeResponse {
+
+}
+
+message UnsubscribeRequest {
+  SubscriberIdentity subscriber = 1;
+}
+
+message UnsubscribeResponse {
+
+}
+
+message PublishRequest {
+  bytes data = 1;
+}
+
+message PublishResponse {}
+
+//this goes out to the delivery actors
+message DeliveryBatch {
+  Subscribers subscribers = 1;
+  ProducerBatch batch = 2;
+}
+
+//messages going into the topic actors
+message ProducerBatch {
+  repeated string type_names = 1;
+  repeated ProducerEnvelope envelopes = 2;
+}
+
+message ProducerEnvelope {
+  int32 type_id = 1;
+  bytes message_data = 2;
+  int32 serializer_id = 3;
+}
+
+//messages going out of the topic actors
+message TopicBatchRequest {
+  repeated string type_names = 1;
+  repeated TopicEnvelope envelopes = 2;
+}
+
+message TopicBatchResponse {
+
+}
+
+message TopicEnvelope {
+  int32 type_id = 1;
+  bytes message_data = 2;
+  int32 serializer_id = 3;
+}
